Add Interactor test
