+++ Songhound.xcodeproj/xcuserdata/DDube.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -47,6 +47,13 @@
 			<key>orderHint</key>
 			<integer>20</integer>
 		</dict>
+		<key>SongsListInteractorTest.xcscheme</key>
+		<dict>
+			<key>isShown</key>
+			<false />
+			<key>orderHint</key>
+			<integer>21</integer>
+		</dict>
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
 	<dict>
